feat: finally, finally, rust-embed and axum-embed work
