+++ backend/src/bin/deploy.rs
@@ -1,6 +1,28 @@
 use backend_lib::schemas::CameraId;
 use backend_lib::pipeline::Pipeline;
-use backend_lib::server::{create_deploy_app, start_server};
+use backend_lib::server::{create_api_router, start_server};
+
+use axum::Router;
+use axum_embed::ServeEmbed;
+use rust_embed::Embed;
+
+// The env var `EMBEDDED_FRONTEND_DIR` is where Bazel placed the frontend
+// static exports, so rust_embed can embed those into this binary.
+#[derive(Embed, Clone)]
+#[folder = "${EMBEDDED_FRONTEND_DIR}"]
+pub struct Asset;
+
+// In deploy mode, the backend will serve the API but instead of serving
+// the Next.js server, it will embed the frontend's static exports using
+// rust-embed and serve using axum_embed.
+pub fn create_deploy_app() -> Router {
+    let serve_assets = ServeEmbed::<Asset>::new();
+        
+    // Use the fallback service so any request that isn't one of the
+    // API's routes will be directed to the frontend static exports.
+    create_api_router()
+        .fallback_service(serve_assets)
+}
 
 // Start tokio async runtime.
 #[tokio::main]
