+++ backend/src/server/mod.rs
@@ -1,6 +1,5 @@
 pub mod app_state;
 pub mod server;
 
-pub use server::create_deploy_app;
-pub use server::create_dev_app;
+pub use server::create_api_router;
 pub use server::start_server;
