+++ frontend/BUILD.bazel
@@ -112,7 +112,7 @@ js_run_binary(
         "package.json",
         "tsconfig.json",
     ],
-    out_dirs = [".next", "out"],
+    out_dirs = ["out"],
     copy_srcs_to_bin = True,
     chdir = package_name(),
     # Environment variable so frontend knows to make requests to same-origin
